replace ExpectImpl.feature with new AnyDomain.featureExtractor
moreover:
- remove NewFeatureAssertionsBuilder
