+++ domain/robstoll-lib/atrium-domain-robstoll-lib-common/src/main/kotlin/ch/tutteli/atrium/domain/robstoll/lib/creating/listAssertions.kt
@@ -5,14 +5,15 @@ package ch.tutteli.atrium.domain.robstoll.lib.creating
 
 import ch.tutteli.atrium.core.Option
 import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.domain.builders.ExpectImpl
+import ch.tutteli.atrium.domain.builders.creating._domain
+import ch.tutteli.atrium.domain.builders.creating.featureExtractor
 import ch.tutteli.atrium.domain.creating.changers.ExtractedFeaturePostStep
 import ch.tutteli.atrium.translations.DescriptionListAssertion
 import kotlin.jvm.JvmMultifileClass
 import kotlin.jvm.JvmName
 
 fun <E, T : List<E>> _get(assertionContainer: Expect<T>, index: Int): ExtractedFeaturePostStep<T, E> =
-    ExpectImpl.feature.extractor(assertionContainer)
+    assertionContainer._domain.featureExtractor
         .methodCall("get", index)
         .withRepresentationForFailure(DescriptionListAssertion.INDEX_OUT_OF_BOUNDS)
         .withFeatureExtraction {
