+++ domain/robstoll-lib/atrium-domain-robstoll-lib-common/src/main/kotlin/ch/tutteli/atrium/domain/robstoll/lib/creating/mapEntryAssertions.kt
@@ -4,6 +4,7 @@ import ch.tutteli.atrium.api.fluent.en_GB.toBe
 import ch.tutteli.atrium.assertions.Assertion
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.domain.builders.ExpectImpl
+import ch.tutteli.atrium.domain.builders.creating._domain
 import ch.tutteli.atrium.domain.builders.creating._domainNullable
 import ch.tutteli.atrium.domain.creating.changers.ExtractedFeaturePostStep
 import kotlin.reflect.KClass
@@ -31,9 +32,9 @@ fun <K : Any, V : Any, T : Map.Entry<K?, V?>> _isKeyValue(
     }
 
 fun <K, T : Map.Entry<K, *>> _key(assertionContainer: Expect<T>): ExtractedFeaturePostStep<T, K> =
-    ExpectImpl.feature.property(assertionContainer, Map.Entry<K, *>::key)
+    assertionContainer._domain.property(Map.Entry<K, *>::key)
 
 fun <V, T : Map.Entry<*, V>> _value(assertionContainer: Expect<T>): ExtractedFeaturePostStep<T, V> =
-    ExpectImpl.feature.property(assertionContainer, Map.Entry<*, V>::value)
+    assertionContainer._domain.property(Map.Entry<*, V>::value)
